# playwright/support/zod/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restful Booker Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Restful Booker Platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/branding": {
      "get": {
        "summary": "Get branding settings",
        "tags": [
          "Branding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Branding settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "hotelName": {
                          "type": "string"
                        },
                        "logo": {
                          "type": "string"
                        },
                        "primaryColor": {
                          "type": "string"
                        },
                        "secondaryColor": {
                          "type": "string"
                        },
                        "fontFamily": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update branding settings (admin only)",
        "tags": [
          "Branding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hotelName": {
                    "type": "string",
                    "example": "Grand Hotel"
                  },
                  "logo": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                  },
                  "primaryColor": {
                    "type": "string",
                    "example": "#FF5733"
                  },
                  "secondaryColor": {
                    "type": "string",
                    "example": "#33FF57"
                  },
                  "fontFamily": {
                    "type": "string",
                    "example": "Arial, sans-serif"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branding settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "hotelName": {
                          "type": "string"
                        },
                        "logo": {
                          "type": "string"
                        },
                        "primaryColor": {
                          "type": "string"
                        },
                        "secondaryColor": {
                          "type": "string"
                        },
                        "fontFamily": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
    "/api/branding/reset": {
      "post": {
        "summary": "Reset branding settings to default (admin only)",
        "tags": [
          "Branding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Branding settings reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "hotelName": {
                          "type": "string"
                        },
                        "logo": {
                          "type": "string"
                        },
                        "primaryColor": {
                          "type": "string"
                        },
                        "secondaryColor": {
                          "type": "string"
                        },
                        "fontFamily": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
...
  },
...
}